+++ packages/rpc-provider/src/coder/encodeObject.spec.ts
@@ -10,19 +10,14 @@ describe('encodeObject', (): void => {
     coder = new RpcCoder();
   });
 
-  it('starts with id === 0 (nothing sent)', (): void => {
-    expect(coder.getId()).toEqual(0);
-  });
-
   it('encodes a valid JsonRPC object', (): void => {
     expect(
       coder.encodeObject('method', ['a', 'b'])
-    ).toEqual({
+    ).toEqual([1, {
       id: 1,
       jsonrpc: '2.0',
       method: 'method',
       params: ['a', 'b']
-    });
-    expect(coder.getId()).toEqual(1);
+    }]);
   });
 });